PromptOptimizerContainer.jsx:
7 Specialized Hooks
Main Orchestrator Simplified
