+++ client/src/features/prompt-optimizer/PromptOptimizerContainer.jsx
@@ -1,716 +1,18 @@
 /**
- * PromptOptimizerContainer V2 - Refactored & Decoupled
- *
- * Reduced from 1,403 lines to ~400 lines by:
- * - Using PromptStateContext for state management
- * - Extracting UI sections into focused components
- * - Delegating responsibilities to smaller components
- *
- * This component now focuses on:
- * - Business logic orchestration
- * - Event handlers
- * - Side effects (loading from URL, highlights persistence, etc.)
+ * BACKWARD COMPATIBILITY SHIM
+ * 
+ * This file has been refactored into a folder structure.
+ * The original implementation has been split into:
+ * - hooks/usePromptLoader.js
+ * - hooks/useHighlightsPersistence.js
+ * - hooks/useUndoRedo.js
+ * - hooks/usePromptOptimization.js
+ * - hooks/useImprovementFlow.js
+ * - hooks/useConceptBrainstorm.js
+ * - hooks/useEnhancementSuggestions.js
+ * - PromptOptimizerContainer/PromptOptimizerContainer.jsx (main orchestrator)
+ * 
+ * This shim maintains backward compatibility with existing imports.
  */
 
-import React, { useEffect, useCallback, useMemo, useRef } from 'react';
-import { usePromptState, PromptStateProvider } from './context/PromptStateContext';
-import { PromptInputSection } from './components/PromptInputSection';
-import { PromptResultsSection } from './components/PromptResultsSection';
-import { PromptModals } from './components/PromptModals';
-import { PromptTopBar } from './components/PromptTopBar';
-import { PromptSidebar } from './components/PromptSidebar';
-import DebugButton from '../../components/DebugButton';
-import { useToast } from '../../components/Toast';
-import { useKeyboardShortcuts } from '../../components/KeyboardShortcuts';
-import { getAuthRepository, getPromptRepository } from '../../repositories';
-import { PromptContext } from '../../utils/PromptContext';
-import { detectAndApplySceneChange } from '../../utils/detectSceneChange';
-import { applySuggestionToPrompt } from './utils/applySuggestion.js';
-import { createHighlightSignature } from './hooks/useSpanLabeling.js';
-import { PERFORMANCE_CONFIG } from '../../config/performance.config';
-import { fetchEnhancementSuggestions as fetchSuggestionsAPI } from './api/enhancementSuggestionsApi';
-
-/**
- * Inner component with access to PromptStateContext
- */
-function PromptOptimizerContent({ user }) {
-  // Force light mode immediately
-  React.useEffect(() => {
-    document.documentElement.classList.remove('dark');
-  }, []);
-
-  const toast = useToast();
-  const {
-    // State
-    selectedMode,
-    showResults,
-    showSettings,
-    setShowSettings,
-    showShortcuts,
-    setShowShortcuts,
-    showHistory,
-    setShowHistory,
-    showImprover,
-    setShowImprover,
-    showBrainstorm,
-    setShowBrainstorm,
-    suggestionsData,
-    setSuggestionsData,
-    setConceptElements,
-    promptContext,
-    setPromptContext,
-    currentPromptUuid,
-    currentPromptDocId,
-    setCurrentPromptUuid,
-    setCurrentPromptDocId,
-    setShowResults,
-
-    // Refs
-    latestHighlightRef,
-    persistedSignatureRef,
-    undoStackRef,
-    redoStackRef,
-    isApplyingHistoryRef,
-    skipLoadFromUrlRef,
-
-    // Hooks
-    promptOptimizer,
-    promptHistory,
-
-    // Functions
-    applyInitialHighlightSnapshot,
-    resetEditStacks,
-    setDisplayedPromptSilently,
-    handleCreateNew,
-
-    // Navigation
-    navigate,
-    uuid,
-  } = usePromptState();
-
-  const aiNames = ['Claude AI', 'ChatGPT', 'Gemini'];
-
-  // Stabilize promptContext to prevent infinite loops
-  const stablePromptContext = useMemo(() => {
-    if (!promptContext) return null;
-    return promptContext;
-  }, [
-    promptContext?.elements?.subject,
-    promptContext?.elements?.action,
-    promptContext?.elements?.location,
-    promptContext?.elements?.time,
-    promptContext?.elements?.mood,
-    promptContext?.elements?.style,
-    promptContext?.elements?.event,
-    promptContext?.metadata?.format,
-    promptContext?.version,
-  ]);
-
-  // ============================================================================
-  // Load prompt from URL parameter
-  // ============================================================================
-  useEffect(() => {
-    const loadPromptFromUrl = async () => {
-      if (!uuid) return;
-      if (skipLoadFromUrlRef.current || currentPromptUuid === uuid) return;
-
-      try {
-        const promptRepository = getPromptRepository();
-        const promptData = await promptRepository.getByUuid(uuid);
-        if (promptData) {
-          promptOptimizer.setInputPrompt(promptData.input);
-          promptOptimizer.setOptimizedPrompt(promptData.output);
-          setDisplayedPromptSilently(promptData.output);
-          setCurrentPromptUuid(promptData.uuid);
-          setCurrentPromptDocId(promptData.id || null);
-          setShowResults(true);
-
-          const preloadHighlight = promptData.highlightCache
-            ? {
-                ...promptData.highlightCache,
-                signature: promptData.highlightCache.signature ?? createHighlightSignature(promptData.output ?? ''),
-              }
-            : null;
-          applyInitialHighlightSnapshot(preloadHighlight, { bumpVersion: true, markPersisted: true });
-          resetEditStacks();
-
-          if (promptData.brainstormContext) {
-            try {
-              const contextData =
-                typeof promptData.brainstormContext === 'string'
-                  ? JSON.parse(promptData.brainstormContext)
-                  : promptData.brainstormContext;
-              const restoredContext = PromptContext.fromJSON(contextData);
-              setPromptContext(restoredContext);
-            } catch (contextError) {
-              console.error('Failed to restore prompt context from shared link:', contextError);
-              toast.warning('Could not restore video context. The prompt will still load.');
-              setPromptContext(null);
-            }
-          } else {
-            setPromptContext(null);
-          }
-        } else {
-          toast.error('Prompt not found');
-          navigate('/', { replace: true });
-        }
-      } catch (error) {
-        console.error('Error loading prompt from URL:', error);
-        toast.error('Failed to load prompt');
-        navigate('/', { replace: true });
-      }
-    };
-
-    loadPromptFromUrl();
-  }, [
-    uuid,
-    currentPromptUuid,
-    navigate,
-    toast,
-    promptOptimizer,
-    setDisplayedPromptSilently,
-    applyInitialHighlightSnapshot,
-    resetEditStacks,
-    setCurrentPromptDocId,
-    setCurrentPromptUuid,
-    setShowResults,
-    setPromptContext,
-    skipLoadFromUrlRef,
-  ]);
-
-  // ============================================================================
-  // Highlights Persistence
-  // ============================================================================
-  const handleHighlightsPersist = useCallback(async (result) => {
-    if (!result || !Array.isArray(result.spans) || !result.signature) {
-      return;
-    }
-
-    const snapshot = {
-      spans: result.spans,
-      meta: result.meta ?? null,
-      signature: result.signature,
-      cacheId: result.cacheId ?? (currentPromptUuid ? String(currentPromptUuid) : null),
-      updatedAt: new Date().toISOString(),
-    };
-
-    const activeCacheId = currentPromptUuid ? String(currentPromptUuid) : null;
-    if (activeCacheId && snapshot.cacheId && snapshot.cacheId !== activeCacheId) {
-      return;
-    }
-
-    latestHighlightRef.current = snapshot;
-    applyInitialHighlightSnapshot(snapshot, { bumpVersion: false, markPersisted: false });
-
-    if (!currentPromptUuid) {
-      return;
-    }
-
-    if (result.source === 'network' || result.source === 'cache-fallback') {
-      promptHistory.updateEntryHighlight(currentPromptUuid, snapshot);
-    }
-
-    if (!user || !currentPromptDocId || result.source !== 'network') {
-      return;
-    }
-
-    if (persistedSignatureRef.current === result.signature) {
-      return;
-    }
-
-    try {
-      const promptRepository = getPromptRepository();
-      await promptRepository.updateHighlights(currentPromptDocId, {
-        highlightCache: snapshot,
-        versionEntry: {
-          versionId: `v-${Date.now()}`,
-          signature: result.signature,
-          spansCount: result.spans.length,
-          timestamp: new Date().toISOString(),
-        },
-      });
-      persistedSignatureRef.current = result.signature;
-    } catch (error) {
-      console.error('Failed to persist highlight snapshot:', error);
-      // Silent failure for background highlight persistence - not critical to user workflow
-      // Only show error if it's a permission issue
-      if (error.code === 'permission-denied') {
-        toast.warning('Unable to save highlights. You may need to sign in.');
-      }
-    }
-  }, [applyInitialHighlightSnapshot, currentPromptDocId, currentPromptUuid, promptHistory, user, latestHighlightRef, persistedSignatureRef]);
-
-  // ============================================================================
-  // Displayed Prompt Changes (with undo/redo support)
-  // ============================================================================
-  const handleDisplayedPromptChange = useCallback(
-    (newText) => {
-      const currentText = promptOptimizer.displayedPrompt;
-      if (isApplyingHistoryRef.current) {
-        isApplyingHistoryRef.current = false;
-        promptOptimizer.setDisplayedPrompt(newText);
-        return;
-      }
-
-      if (currentText !== newText) {
-        undoStackRef.current = [...undoStackRef.current, {
-          text: currentText,
-          highlight: latestHighlightRef.current,
-        }].slice(-PERFORMANCE_CONFIG.UNDO_STACK_SIZE);
-        redoStackRef.current = [];
-      }
-
-      promptOptimizer.setDisplayedPrompt(newText);
-    },
-    [promptOptimizer, isApplyingHistoryRef, undoStackRef, redoStackRef, latestHighlightRef]
-  );
-
-  // Track setTimeout IDs for cleanup
-  const undoTimeoutRef = useRef(null);
-  const redoTimeoutRef = useRef(null);
-  const conceptOptimizeTimeoutRef = useRef(null);
-
-  // Cleanup timeouts on unmount
-  useEffect(() => {
-    return () => {
-      if (undoTimeoutRef.current) clearTimeout(undoTimeoutRef.current);
-      if (redoTimeoutRef.current) clearTimeout(redoTimeoutRef.current);
-      if (conceptOptimizeTimeoutRef.current) clearTimeout(conceptOptimizeTimeoutRef.current);
-    };
-  }, []);
-
-  // ============================================================================
-  // Undo/Redo Handlers
-  // ============================================================================
-  const handleUndo = useCallback(() => {
-    if (!undoStackRef.current.length) return;
-
-    const previous = undoStackRef.current.pop();
-    const currentSnapshot = {
-      text: promptOptimizer.displayedPrompt,
-      highlight: latestHighlightRef.current,
-    };
-    redoStackRef.current = [...redoStackRef.current, currentSnapshot].slice(-PERFORMANCE_CONFIG.UNDO_STACK_SIZE);
-
-    isApplyingHistoryRef.current = true;
-    setDisplayedPromptSilently(previous.text);
-    promptOptimizer.setOptimizedPrompt(previous.text);
-    applyInitialHighlightSnapshot(previous.highlight ?? null, { bumpVersion: true, markPersisted: false });
-
-    // Clear any existing timeout and set new one
-    if (undoTimeoutRef.current) clearTimeout(undoTimeoutRef.current);
-    undoTimeoutRef.current = setTimeout(() => {
-      isApplyingHistoryRef.current = false;
-      undoTimeoutRef.current = null;
-    }, PERFORMANCE_CONFIG.REF_RESET_DELAY_MS);
-  }, [applyInitialHighlightSnapshot, promptOptimizer, setDisplayedPromptSilently, undoStackRef, redoStackRef, latestHighlightRef, isApplyingHistoryRef]);
-
-  const handleRedo = useCallback(() => {
-    if (!redoStackRef.current.length) return;
-
-    const next = redoStackRef.current.pop();
-    undoStackRef.current = [...undoStackRef.current, {
-      text: promptOptimizer.displayedPrompt,
-      highlight: latestHighlightRef.current,
-    }].slice(-PERFORMANCE_CONFIG.UNDO_STACK_SIZE);
-
-    isApplyingHistoryRef.current = true;
-    setDisplayedPromptSilently(next.text);
-    promptOptimizer.setOptimizedPrompt(next.text);
-    applyInitialHighlightSnapshot(next.highlight ?? null, { bumpVersion: true, markPersisted: false });
-
-    // Clear any existing timeout and set new one
-    if (redoTimeoutRef.current) clearTimeout(redoTimeoutRef.current);
-    redoTimeoutRef.current = setTimeout(() => {
-      isApplyingHistoryRef.current = false;
-      redoTimeoutRef.current = null;
-    }, PERFORMANCE_CONFIG.REF_RESET_DELAY_MS);
-  }, [applyInitialHighlightSnapshot, promptOptimizer, setDisplayedPromptSilently, undoStackRef, redoStackRef, latestHighlightRef, isApplyingHistoryRef]);
-
-  // ============================================================================
-  // Optimization Handler
-  // ============================================================================
-  const handleOptimize = async (promptToOptimize, context) => {
-    const prompt = promptToOptimize || promptOptimizer.inputPrompt;
-    const ctx = context || promptOptimizer.improvementContext;
-
-    const serializedContext = promptContext
-      ? typeof promptContext.toJSON === 'function'
-        ? promptContext.toJSON()
-        : {
-            elements: promptContext.elements,
-            metadata: promptContext.metadata,
-          }
-      : null;
-
-    const brainstormContextData = serializedContext
-      ? {
-          elements: serializedContext.elements,
-          metadata: serializedContext.metadata,
-        }
-      : null;
-
-    const result = await promptOptimizer.optimize(prompt, ctx, brainstormContextData);
-    if (result) {
-      const saveResult = await promptHistory.saveToHistory(
-        prompt,
-        result.optimized,
-        result.score,
-        selectedMode,
-        serializedContext
-      );
-
-      if (saveResult?.uuid) {
-        skipLoadFromUrlRef.current = true;
-        setCurrentPromptUuid(saveResult.uuid);
-        setCurrentPromptDocId(saveResult.id ?? null);
-        setDisplayedPromptSilently(result.optimized);
-        setShowResults(true);
-        applyInitialHighlightSnapshot(null, { bumpVersion: true, markPersisted: false });
-        resetEditStacks();
-        persistedSignatureRef.current = null;
-        if (saveResult.uuid) {
-          navigate(`/prompt/${saveResult.uuid}`, { replace: true });
-        }
-      }
-    }
-  };
-
-  // ============================================================================
-  // Improvement Flow
-  // ============================================================================
-  const handleImproveFirst = () => {
-    if (!promptOptimizer.inputPrompt.trim()) {
-      toast.warning('Please enter a prompt first');
-      return;
-    }
-    setShowImprover(true);
-  };
-
-  const handleImprovementComplete = async (enhancedPrompt, context) => {
-    setShowImprover(false);
-    promptOptimizer.setImprovementContext(context);
-    promptOptimizer.setInputPrompt(enhancedPrompt);
-    handleOptimize(enhancedPrompt, context);
-  };
-
-  // ============================================================================
-  // Brainstorm/Video Concept Flow
-  // ============================================================================
-  const handleConceptComplete = async (finalConcept, elements, metadata) => {
-    setConceptElements(elements);
-
-    const context = new PromptContext(elements, metadata);
-    setPromptContext(context);
-
-    const serializedContext = context.toJSON();
-    const brainstormContextData = {
-      elements,
-      metadata
-    };
-
-    promptOptimizer.setInputPrompt(finalConcept);
-    setShowBrainstorm(false);
-
-    // Clear any existing timeout and set new one
-    if (conceptOptimizeTimeoutRef.current) clearTimeout(conceptOptimizeTimeoutRef.current);
-    conceptOptimizeTimeoutRef.current = setTimeout(async () => {
-      try {
-        const result = await promptOptimizer.optimize(finalConcept, null, brainstormContextData);
-        if (result) {
-          const saveResult = await promptHistory.saveToHistory(
-            finalConcept,
-            result.optimized,
-            result.score,
-            selectedMode,
-            serializedContext
-          );
-          if (saveResult?.uuid) {
-            setDisplayedPromptSilently(result.optimized);
-
-            skipLoadFromUrlRef.current = true;
-            setCurrentPromptUuid(saveResult.uuid);
-            setCurrentPromptDocId(saveResult.id ?? null);
-            setShowResults(true);
-            toast.success('Video prompt generated successfully!');
-            applyInitialHighlightSnapshot(null, { bumpVersion: true, markPersisted: false });
-            resetEditStacks();
-            persistedSignatureRef.current = null;
-            navigate(`/prompt/${saveResult.uuid}`, { replace: true });
-          }
-        }
-      } catch (error) {
-        toast.error('Failed to generate video prompt. Please try again.');
-        console.error('Error in handleConceptComplete:', error);
-      } finally {
-        conceptOptimizeTimeoutRef.current = null;
-      }
-    }, PERFORMANCE_CONFIG.ASYNC_OPERATION_DELAY_MS);
-  };
-
-  const handleSkipBrainstorm = () => {
-    setShowBrainstorm(false);
-    setConceptElements({ skipped: true });
-  };
-
-  // ============================================================================
-  // Enhancement Suggestions
-  // ============================================================================
-  const handleSuggestionClick = async (suggestion) => {
-    const suggestionText =
-      typeof suggestion === 'string' ? suggestion : suggestion?.text || '';
-
-    if (!suggestionText || !suggestionsData) return;
-
-    const {
-      selectedText,
-      fullPrompt,
-      range,
-      offsets,
-      metadata,
-    } = suggestionsData;
-
-    try {
-      // Apply the suggestion (delegates to utility)
-      const result = await applySuggestionToPrompt({
-        currentPrompt: fullPrompt,
-        selectedText,
-        suggestionText,
-        range,
-        offsets,
-        metadata,
-      });
-
-      // Update displayed prompt
-      if (result.updatedPrompt) {
-        setDisplayedPromptSilently(result.updatedPrompt);
-        toast.success('Suggestion applied');
-      }
-
-      // Close suggestions panel
-      setSuggestionsData(null);
-    } catch (error) {
-      console.error('Error applying suggestion:', error);
-      toast.error('Failed to apply suggestion');
-    }
-  };
-
-  const fetchEnhancementSuggestions = async (payload = {}) => {
-    const {
-      highlightedText,
-      originalText,
-      displayedPrompt: payloadPrompt,
-      range,
-      offsets,
-      metadata: rawMetadata = null,
-      trigger = 'highlight',
-    } = payload;
-
-    const trimmedHighlight = (highlightedText || '').trim();
-    const rawPrompt = payloadPrompt ?? promptOptimizer.displayedPrompt ?? '';
-    const normalizedPrompt = rawPrompt.normalize('NFC');
-    const metadata = rawMetadata
-      ? {
-          ...rawMetadata,
-          span: rawMetadata.span ? { ...rawMetadata.span } : null,
-        }
-      : null;
-
-    // Early returns for invalid cases
-    if (selectedMode !== 'video' || !trimmedHighlight) {
-      return;
-    }
-
-    if (suggestionsData?.selectedText === trimmedHighlight && suggestionsData?.show) {
-      return;
-    }
-
-    // Show loading state immediately
-    setSuggestionsData({
-      show: true,
-      selectedText: trimmedHighlight,
-      originalText: originalText || trimmedHighlight,
-      suggestions: [],
-      isLoading: true,
-      isPlaceholder: false,
-      fullPrompt: normalizedPrompt,
-      range,
-      offsets,
-      metadata,
-      setSuggestions: (newSuggestions, newIsPlaceholder) => {
-        setSuggestionsData((prev) => ({
-          ...prev,
-          suggestions: newSuggestions,
-          isPlaceholder:
-            newIsPlaceholder !== undefined
-              ? newIsPlaceholder
-              : prev.isPlaceholder,
-        }));
-      },
-      onSuggestionClick: handleSuggestionClick,
-      onClose: () => setSuggestionsData(null),
-    });
-
-    try {
-      // Delegate to API layer (VideoConceptBuilder pattern)
-      const { suggestions, isPlaceholder } = await fetchSuggestionsAPI({
-        highlightedText: trimmedHighlight,
-        normalizedPrompt,
-        inputPrompt: promptOptimizer.inputPrompt,
-        brainstormContext: stablePromptContext,
-        metadata,
-      });
-
-      // Update with results
-      setSuggestionsData(prev => ({
-        ...prev,
-        suggestions,
-        isLoading: false,
-        isPlaceholder,
-      }));
-    } catch (error) {
-      console.error('Error fetching suggestions:', error);
-      toast.error('Failed to load suggestions');
-
-      setSuggestionsData((prev) => ({
-        ...prev,
-        isLoading: false,
-        suggestions: [{ text: 'Failed to load suggestions. Please try again.' }],
-      }));
-    }
-  };
-
-  // ============================================================================
-  // Keyboard Shortcuts
-  // ============================================================================
-  useKeyboardShortcuts({
-    openShortcuts: () => setShowShortcuts(true),
-    openSettings: () => setShowSettings(true),
-    createNew: handleCreateNew,
-    optimize: () => !promptOptimizer.isProcessing && showResults === false && handleOptimize(),
-    improveFirst: handleImproveFirst,
-    canCopy: () => showResults && promptOptimizer.displayedPrompt,
-    copy: () => {
-      navigator.clipboard.writeText(promptOptimizer.displayedPrompt);
-      toast.success('Copied to clipboard!');
-    },
-    export: () => showResults && toast.info('Use export button in canvas'),
-    toggleSidebar: () => setShowHistory(!showHistory),
-    switchMode: (index) => {
-      // Implementation from original
-    },
-    applySuggestion: (index) => {
-      if (suggestionsData?.suggestions[index]) {
-        suggestionsData.onSuggestionClick(suggestionsData.suggestions[index]);
-      }
-    },
-    closeModal: () => {
-      if (showSettings) setShowSettings(false);
-      else if (showShortcuts) setShowShortcuts(false);
-      else if (showImprover) setShowImprover(false);
-      else if (showBrainstorm) setShowBrainstorm(false);
-      else if (suggestionsData) setSuggestionsData(null);
-    },
-  });
-
-  // ============================================================================
-  // Render
-  // ============================================================================
-  return (
-    <div
-      className="h-screen overflow-hidden gradient-neutral transition-colors duration-300"
-      style={{ '--sidebar-width': showHistory ? '18rem' : '0px' }}
-    >
-      {/* Skip to main content */}
-      <a href="#main-content" className="sr-only-focusable top-4 left-4">
-        Skip to main content
-      </a>
-
-      {/* Modals */}
-      <PromptModals
-        onImprovementComplete={handleImprovementComplete}
-        onConceptComplete={handleConceptComplete}
-        onSkipBrainstorm={handleSkipBrainstorm}
-      />
-
-      {/* Top Action Buttons */}
-      <PromptTopBar />
-
-      {/* History Sidebar */}
-      <PromptSidebar user={user} />
-
-      {/* Main Content */}
-      <main
-        id="main-content"
-        className={`relative flex h-screen flex-col items-center px-4 sm:px-6 py-8 transition-all duration-300 ${showHistory ? 'ml-72' : 'ml-0'} ${showResults ? 'justify-start' : 'justify-center overflow-y-auto'}`}
-      >
-        {/* Input Section */}
-        {!showResults && (
-          <PromptInputSection
-            aiNames={aiNames}
-            onOptimize={handleOptimize}
-            onShowBrainstorm={() => setShowBrainstorm(true)}
-          />
-        )}
-
-        {/* Results Section */}
-        <PromptResultsSection
-          onDisplayedPromptChange={handleDisplayedPromptChange}
-          onFetchSuggestions={fetchEnhancementSuggestions}
-          onHighlightsPersist={handleHighlightsPersist}
-          onUndo={handleUndo}
-          onRedo={handleRedo}
-          stablePromptContext={stablePromptContext}
-        />
-
-        {/* Privacy Policy Footer */}
-        {!showResults && (
-          <footer className="absolute bottom-8 left-1/2 transform -translate-x-1/2">
-            <a
-              href="/privacy-policy"
-              className="text-sm text-neutral-500 hover:text-neutral-700 transition-colors"
-            >
-              Privacy Policy
-            </a>
-          </footer>
-        )}
-      </main>
-
-      {/* Debug Button */}
-      {(process.env.NODE_ENV === 'development' ||
-        new URLSearchParams(window.location.search).get('debug') === 'true') && (
-        <DebugButton
-          inputPrompt={promptOptimizer.inputPrompt}
-          displayedPrompt={promptOptimizer.displayedPrompt}
-          optimizedPrompt={promptOptimizer.optimizedPrompt}
-          selectedMode={selectedMode}
-          promptContext={stablePromptContext}
-        />
-      )}
-    </div>
-  );
-}
-
-/**
- * Outer component with auth state management
- */
-function PromptOptimizerContainer() {
-  const [user, setUser] = React.useState(null);
-
-  // Listen for auth state changes
-  React.useEffect(() => {
-    const authRepository = getAuthRepository();
-    const unsubscribe = authRepository.onAuthStateChanged((currentUser) => {
-      setUser(currentUser);
-    });
-    return () => unsubscribe();
-  }, []);
-
-  return (
-    <PromptStateProvider user={user}>
-      <PromptOptimizerContent user={user} />
-    </PromptStateProvider>
-  );
-}
-
-export default PromptOptimizerContainer;
+export { default } from './PromptOptimizerContainer/PromptOptimizerContainer.jsx';
