+++ client/src/features/prompt-optimizer/PromptOptimizerContainer/hooks/index.js
@@ -1,0 +1,13 @@
+/**
+ * PromptOptimizerContainer Hooks - Barrel Exports
+ * Provides clean imports for all custom hooks
+ */
+
+export { usePromptLoader } from './usePromptLoader';
+export { useHighlightsPersistence } from './useHighlightsPersistence';
+export { useUndoRedo } from './useUndoRedo';
+export { usePromptOptimization } from './usePromptOptimization';
+export { useImprovementFlow } from './useImprovementFlow';
+export { useConceptBrainstorm } from './useConceptBrainstorm';
+export { useEnhancementSuggestions } from './useEnhancementSuggestions';
+
